// frontend/views/js/src/pages/user/data/models/UserModel.ts
export interface UserModel {
    UUID: string;
    username: string;
    fullName: string;
    profileImage: string;
    email: string;
    description: string;
    address: string;
    gender: string;
    title: string;
    phoneNo: string;
    valid: boolean;
    createdAtFormat: string;
    createdByName: string;
    updatedAtFormat: string;
    updatedByName: string;
}